Allow Admin to change logos (#203) (#221)
Allow ISVs to upload logos using the admin page.  Logos need to be uploaded in .png for the top-left icon and .ico for the favicon tab icon.  After the upload is uploaded then a restart of the webapps will be required.  Browser cache will keep the old logo after restart so we recommend ctrl+f5 to force refresh.